refactor(sync-service): Split syncBatch routes out of controller
Moves route/middleware declarations and response Zod schemas into
syncBatch.routes.ts; controller keeps only request-handling logic via
createSyncBatchController(service). module.ts now creates the shared
DocumentedRouter and registers routes into it.

No behavior change — verified route-for-route identical OpenAPI doc
before/after, lint clean, all 4 existing tests pass, and live smoke
checks against the running service.
